Add files via upload
Analyzed the results of different measurements Networkx offers, for example, Degree Centrality, degree_in and degree_out centrality (for the whole data-set, for the daytime and for nighttime),  closeness centrality, betweenness centrality. The idea was just to get e bettwen understandings of the nodes/regions and study which regions are most important to network.